Updated Timesless Quiz
I have provided the styling code for the 'LINGO' within this component. The original code has been kept and commented out above the new h1 elemental change.

It is my suggestion that we find a way of duplicating the styling for 'LINGO' here so that when the screen is in light mode, the 'O' in lingo turns black in colour, and white colour when in dark mode!, This would enhance the styling and keep the text from blending into the background.